refactor(payment): move payment handlers to handler/payment
- Moved DepositPersistenceHandler to the payment package and created PaymentCompletedHandler, PaymentIdPersistenceHandler, and PaymentInitiationHandler to manage payment events.
- Added tests for PaymentCompletedHandler and PaymentIdPersistenceHandler to ensure correct functionality.
- Updated webhook integration to utilize new payment handlers, enhancing the event-driven architecture for payment processing.
- Introduced a mock payment provider for testing payment initiation logic.
