+++ pkg/handler/payment/completed_handler_test.go
@@ -1,0 +1,8 @@
+// Stub for PaymentCompletedHandler tests. Add/port tests as needed.
+package payment
+
+import "testing"
+
+func TestPaymentCompletedHandler(t *testing.T) {
+	// TODO: Implement tests
+}
