+++ pkg/handler/payment/failed_handler.go
@@ -1,0 +1,4 @@
+package payment
+
+// PaymentFailedHandler is a placeholder for handling PaymentFailedEvent.
+// Implement logic as needed for your event-driven payment flow.
