+++ pkg/handler/payment/id_persisted_handler_test.go
@@ -1,0 +1,8 @@
+// Stub for PaymentIdPersistenceHandler tests. Add/port tests as needed.
+package payment
+
+import "testing"
+
+func TestPaymentIdPersistenceHandler(t *testing.T) {
+	// TODO: Implement tests
+}
