+++ pkg/handler/payment/initiation_handler.go
@@ -1,4 +1,4 @@
-package deposit
+package payment
 
 import (
 	"context"
