+++ webapi/account/webhook_test.go
@@ -10,13 +10,14 @@ import (
 
 	"log/slog"
 
+	"github.com/amirasaad/fintech/infra/provider"
 	fixturesmocks "github.com/amirasaad/fintech/internal/fixtures/mocks"
 	"github.com/amirasaad/fintech/pkg/domain"
 	accountdomain "github.com/amirasaad/fintech/pkg/domain/account"
 	"github.com/amirasaad/fintech/pkg/domain/account/events"
 	"github.com/amirasaad/fintech/pkg/domain/money"
 	"github.com/amirasaad/fintech/pkg/eventbus"
-	"github.com/amirasaad/fintech/pkg/handler/account/deposit"
+	"github.com/amirasaad/fintech/pkg/handler/payment"
 	"github.com/amirasaad/fintech/webapi/account"
 	"github.com/gofiber/fiber/v2"
 	"github.com/stretchr/testify/assert"
@@ -92,7 +93,9 @@ func TestStripeWebhookHandler_Integration(t *testing.T) {
 	// Set up event bus and register handler
 	bus := eventbus.NewSimpleEventBus()
 	logger := slog.New(slog.NewTextHandler(nil, nil))
-	bus.Subscribe((events.PaymentCompletedEvent{}).EventType(), deposit.PaymentCompletedHandler(bus, uow, logger))
+	bus.Subscribe((events.PaymentCompletedEvent{}).EventType(), payment.PaymentCompletedHandler(bus, uow, logger))
+	bus.Subscribe((events.PaymentInitiationEvent{}).EventType(), payment.PaymentInitiationHandler(bus, provider.NewMockPaymentProvider()))
+	bus.Subscribe((events.PaymentIdPersistedEvent{}).EventType(), payment.PaymentIdPersistenceHandler(bus, uow))
 
 	// Set up Fiber app with webhook handler
 	app := fiber.New()
